Finished exercise 2.15*: The Phonebook step 10 Why is there a star in the exercise? See here for the explanation.
Change the functionality so that if a number is added to an already existing user, the new number will replace the old number. It's recommended to use the HTTP PUT method for updating the phone number.
-------
Modifyed the addNameAndNumber method and handled the put method toward the server in the PersonService.js module