Added logs to help with debugging.
Fixed the code when deleting a user in the interface.
Before: The person data to be removed in the frontend was the person data returned by the server.
Now: The person data to be removed in the frontend is independend from the server backend.
Problem: If server gives 200 OK message for some reason even though it does not delete the entry, the data in the frontend and in the backend will be different.